Create new #preorder which calls the old one
Changed it to able to apply yield to all nodes
that was taken in preoder traversal

The old way is a recursion itself which is unable
to yield the block starting on the second recursio
n. Now we use the recursion mathod to get the
nodes in a preoder and yield each node before
returning them.